+++ src/main/kotlin/org/mybatis/dynamic/sql/util/kotlin/JoinCollector.kt
@@ -17,39 +17,29 @@ package org.mybatis.dynamic.sql.util.kotlin
 
 import org.mybatis.dynamic.sql.BindableColumn
 import org.mybatis.dynamic.sql.SqlBuilder
-import org.mybatis.dynamic.sql.select.join.JoinCondition
-import org.mybatis.dynamic.sql.select.join.JoinCriterion
+import org.mybatis.dynamic.sql.VisitableCondition
 
 typealias JoinReceiver = JoinCollector.() -> Unit
 
 @MyBatisDslMarker
 class JoinCollector {
-    private var onJoinCriterion: JoinCriterion<*>? = null
-    internal val andJoinCriteria = mutableListOf<JoinCriterion<*>>()
+    private val criteriaCollector = GroupingCriteriaCollector()
 
-    internal fun onJoinCriterion() : JoinCriterion<*> = invalidIfNull(onJoinCriterion, "ERROR.22") //$NON-NLS-1$
+    internal fun initialCriterion() = invalidIfNull(criteriaCollector.initialCriterion, "ERROR.22") //$NON-NLS-1$
+    internal fun subCriteria() = criteriaCollector.subCriteria
 
     fun <T> on(leftColumn: BindableColumn<T>): RightColumnCollector<T> = RightColumnCollector {
-        onJoinCriterion = JoinCriterion.Builder<T>()
-            .withConnector("on") //$NON-NLS-1$
-            .withJoinColumn(leftColumn)
-            .withJoinCondition(it)
-            .build()
+        assertNull(criteriaCollector.initialCriterion, "ERROR.45") //$NON-NLS-1$
+        criteriaCollector.apply { leftColumn.invoke(it) }
     }
 
     fun <T> and(leftColumn: BindableColumn<T>): RightColumnCollector<T> = RightColumnCollector {
-        andJoinCriteria.add(
-            JoinCriterion.Builder<T>()
-                .withConnector("and") //$NON-NLS-1$
-                .withJoinColumn(leftColumn)
-                .withJoinCondition(it)
-                .build()
-        )
+        criteriaCollector.and { leftColumn.invoke(it) }
     }
 }
 
-class RightColumnCollector<T>(private val joinConditionConsumer: (JoinCondition<T>) -> Unit) {
-    infix fun equalTo(rightColumn: BindableColumn<T>) = joinConditionConsumer.invoke(SqlBuilder.equalTo(rightColumn))
+class RightColumnCollector<T>(private val joinConditionConsumer: (VisitableCondition<T>) -> Unit) {
+    infix fun equalTo(rightColumn: BindableColumn<T>) = joinConditionConsumer.invoke(SqlBuilder.isEqualTo(rightColumn))
 
-    infix fun equalTo(value: T) = joinConditionConsumer.invoke(SqlBuilder.equalTo(value))
+    infix fun equalTo(value: T) = joinConditionConsumer.invoke(SqlBuilder.isEqualTo(value))
 }