+++ src/test/kotlin/examples/kotlin/mybatis3/joins/JoinMapperNewSyntaxTest.kt
@@ -32,12 +32,11 @@ import org.mybatis.dynamic.sql.util.Messages
 import org.mybatis.dynamic.sql.util.kotlin.KInvalidSQLException
 import org.mybatis.dynamic.sql.util.kotlin.elements.constant
 import org.mybatis.dynamic.sql.util.kotlin.elements.invoke
-import org.mybatis.dynamic.sql.util.kotlin.elements.max
 import org.mybatis.dynamic.sql.util.kotlin.mybatis3.select
 
 @Suppress("LargeClass")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class JoinMapperTest {
+class JoinMapperNewSyntaxTest {
     private lateinit var sqlSessionFactory: SqlSessionFactory
 
     @BeforeAll
@@ -58,8 +57,8 @@ class JoinMapperTest {
                 orderDetail.lineNumber, orderDetail.description, orderDetail.quantity
             ) {
                 from(orderMaster, "om")
-                join(orderDetail, "od") {
-                    on(orderMaster.orderId) equalTo orderDetail.orderId
+                join(orderDetail, "od") on {
+                    orderMaster.orderId isEqualTo orderDetail.orderId
                 }
             }
 
@@ -97,9 +96,9 @@ class JoinMapperTest {
                 orderDetail.lineNumber, orderDetail.description, orderDetail.quantity
             ) {
                 from(orderMaster, "om")
-                join(orderDetail, "od") {
-                    on(orderMaster.orderId) equalTo orderDetail.orderId
-                    and(orderMaster.orderId) equalTo 1
+                join(orderDetail, "od") on {
+                    orderMaster.orderId isEqualTo orderDetail.orderId
+                    and { orderMaster.orderId isEqualTo 1 }
                 }
             }
 
@@ -132,9 +131,9 @@ class JoinMapperTest {
                 orderDetail.lineNumber, orderDetail.description, orderDetail.quantity
             ) {
                 from(orderMaster, "om")
-                join(orderDetail, "od") {
-                    on(orderMaster.orderId) equalTo orderDetail.orderId
-                    and(orderMaster.orderId) equalTo constant("1")
+                join(orderDetail, "od") on {
+                    orderMaster.orderId isEqualTo orderDetail.orderId
+                    and { orderMaster.orderId isEqualTo constant<Int>("1") }
                 }
             }
 
@@ -165,9 +164,9 @@ class JoinMapperTest {
             orderDetail.description, orderDetail.quantity
         ) {
             from(orderMaster, "om")
-            join(orderDetail, "od") {
-                on(orderMaster.orderId) equalTo orderDetail.orderId
-                and(orderMaster.orderId) equalTo orderDetail.orderId
+            join(orderDetail, "od") on {
+                orderMaster.orderId isEqualTo orderDetail.orderId
+                and { orderMaster.orderId isEqualTo orderDetail.orderId }
             }
         }
 
@@ -184,9 +183,9 @@ class JoinMapperTest {
             orderDetail.description, orderDetail.quantity
         ) {
             from(orderMaster, "om")
-            join(orderDetail, "od") {
-                on(orderMaster.orderId) equalTo orderDetail.orderId
-                and(orderMaster.orderId) equalTo orderDetail.orderId
+            join(orderDetail, "od") on {
+                orderMaster.orderId isEqualTo orderDetail.orderId
+                and { orderMaster.orderId isEqualTo orderDetail.orderId }
             }
             where { orderMaster.orderId isEqualTo 1 }
         }
@@ -207,11 +206,11 @@ class JoinMapperTest {
                 itemMaster.description, orderLine.quantity
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                join(itemMaster, "im") {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                join(itemMaster, "im") on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 where { orderMaster.orderId isEqualTo 2 }
             }
@@ -243,11 +242,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                fullJoin(itemMaster, "im") {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                fullJoin(itemMaster, "im") on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -311,28 +310,21 @@ class JoinMapperTest {
                     }
                     + "om"
                 }
-                join(
-                    subQuery = {
-                        select(orderLine.allColumns()) {
-                            from(orderLine)
-                        }
-                        + "ol"
-                    },
-                    joinCriteria = {
-                        on("om"(orderMaster.orderId)) equalTo
-                                "ol"(orderLine.orderId)
+                join {
+                    select(orderLine.allColumns()) {
+                        from(orderLine)
                     }
-                )
-                fullJoin(
-                    {
-                        select(itemMaster.allColumns()) {
-                            from(itemMaster)
-                        }
-                        + "im"
+                    + "ol"
+                } on {
+                    "om"(orderMaster.orderId) isEqualTo "ol"(orderLine.orderId)
+                }
+                fullJoin {
+                    select(itemMaster.allColumns()) {
+                        from(itemMaster)
                     }
-                ) {
-                    on("ol"(orderLine.itemId)) equalTo
-                            "im"(itemMaster.itemId)
+                    +"im"
+                } on {
+                    "ol"(orderLine.itemId) isEqualTo "im"(itemMaster.itemId)
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -390,11 +382,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                fullJoin(itemMaster) {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                fullJoin(itemMaster) on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -438,11 +430,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                leftJoin(itemMaster, "im") {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                leftJoin(itemMaster, "im") on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -482,8 +474,8 @@ class JoinMapperTest {
                 itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
                 leftJoin(
                     {
@@ -492,8 +484,8 @@ class JoinMapperTest {
                         }
                         + "im"
                     }
-                ) {
-                    on(orderLine.itemId) equalTo "im"(itemMaster.itemId)
+                ) on {
+                    orderLine.itemId isEqualTo "im"(itemMaster.itemId)
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -532,11 +524,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                leftJoin(itemMaster) {
-                    on(orderLine.itemId)  equalTo itemMaster.itemId
+                leftJoin(itemMaster) on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -575,11 +567,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                rightJoin(itemMaster, "im") {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                rightJoin(itemMaster, "im") on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -619,8 +611,8 @@ class JoinMapperTest {
                 "im"(itemMaster.itemId), itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
                 rightJoin(
                     {
@@ -629,8 +621,8 @@ class JoinMapperTest {
                         }
                         + "im"
                     }
-                ) {
-                    on(orderLine.itemId) equalTo "im"(itemMaster.itemId)
+                ) on {
+                    orderLine.itemId isEqualTo "im"(itemMaster.itemId)
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -669,11 +661,11 @@ class JoinMapperTest {
                 orderLine.orderId, orderLine.quantity, itemMaster.itemId, itemMaster.description
             ) {
                 from(orderMaster, "om")
-                join(orderLine, "ol") {
-                    on(orderMaster.orderId) equalTo orderLine.orderId
+                join(orderLine, "ol") on {
+                    orderMaster.orderId isEqualTo orderLine.orderId
                 }
-                rightJoin(itemMaster) {
-                    on(orderLine.itemId) equalTo itemMaster.itemId
+                rightJoin(itemMaster) on {
+                    orderLine.itemId isEqualTo itemMaster.itemId
                 }
                 orderBy(orderLine.orderId, itemMaster.itemId)
             }
@@ -714,8 +706,8 @@ class JoinMapperTest {
             // get Bamm Bamm's parent - should be Barney
             val selectStatement = select(user.userId, user.userName, user.parentId) {
                 from(user, "u1")
-                join(user2, "u2") {
-                    on(user.userId) equalTo user2.parentId
+                join(user2, "u2") on {
+                    user.userId isEqualTo user2.parentId
                 }
                 where { user2.userId isEqualTo 4 }
             }
@@ -745,8 +737,8 @@ class JoinMapperTest {
             // get Bamm Bamm's parent - should be Barney
             val selectStatement = select(user.userId, user.userName, user.parentId) {
                 from(user)
-                join(user2) {
-                    on(user.userId) equalTo user2.parentId
+                join(user2) on {
+                    user.userId isEqualTo user2.parentId
                 }
                 where { user2.userId isEqualTo 4 }
             }
@@ -777,8 +769,39 @@ class JoinMapperTest {
             // get Bamm Bamm's parent - should be Barney
             val selectStatement = select(user.userId, user.userName, user.parentId) {
                 from(user, "u1")
-                join(user2, "u2") {
-                    on(user.userId) equalTo user2.parentId
+                join(user2, "u2") on {
+                    user.userId isEqualTo user2.parentId
+                }
+                where { user2.userId isEqualTo 4 }
+            }
+
+            val expectedStatement = "select u1.user_id, u1.user_name, u1.parent_id" +
+                    " from User u1 join User u2 on u1.user_id = u2.parent_id" +
+                    " where u2.user_id = #{parameters.p1,jdbcType=INTEGER}"
+            assertThat(selectStatement.selectStatement).isEqualTo(expectedStatement)
+            val rows = mapper.selectManyMappedRows(selectStatement)
+            assertThat(rows).hasSize(1)
+
+            assertThat(rows[0]).containsExactly(
+                entry("USER_ID", 2),
+                entry("USER_NAME", "Barney"),
+            )
+        }
+    }
+
+    @Test
+    fun testSelfWithNewAliasAndOverrideOddUsage() {
+        sqlSessionFactory.openSession().use { session ->
+            val mapper = session.getMapper(JoinMapper::class.java)
+
+            // create second table instance for self-join
+            val user2 = user.withAlias("other_user")
+
+            // get Bamm Bamm's parent - should be Barney
+            val selectStatement = select(user.userId, user.userName, user.parentId) {
+                from(user, "u1")
+                join(user2, "u2") on {
+                    and { user.userId isEqualTo user2.parentId }
                 }
                 where { user2.userId isEqualTo 4 }
             }
@@ -805,72 +828,9 @@ class JoinMapperTest {
         assertThatExceptionOfType(KInvalidSQLException::class.java).isThrownBy {
             select(user.userId, user.userName, user.parentId) {
                 from(user, "u1")
-                join(user2, "u2") {
-                    and(user.userId) equalTo user2.parentId
-                }
+                join(user2, "u2") on { }
                 where { user2.userId isEqualTo 4 }
             }
         }.withMessage(Messages.getString("ERROR.22")) //$NON-NLS-1$
-    }
-
-    @Test
-    fun testJoinWithDoubleOnCondition() {
-        // create second table instance for self-join
-        val user2 = user.withAlias("other_user")
-
-        assertThatExceptionOfType(KInvalidSQLException::class.java).isThrownBy {
-            select(user.userId, user.userName, user.parentId) {
-                from(user, "u1")
-                join(user2, "u2") {
-                    on(user.userId) equalTo user2.parentId
-                    on(user.userId) equalTo user2.parentId
-                }
-                where { user2.userId isEqualTo 4 }
-            }
-        }.withMessage(Messages.getString("ERROR.45")) //$NON-NLS-1$
-    }
-
-    @Test
-    fun testThatAliasesPropagateToSubQueryConditions() {
-        sqlSessionFactory.openSession().use { session ->
-            val mapper = session.getMapper(JoinMapper::class.java)
-
-            val orderLine2 = OrderLineDynamicSQLSupport.OrderLine()
-
-            val selectStatement = select(orderLine.orderId, orderLine.lineNumber) {
-                from(orderLine, "ol")
-                where {
-                    orderLine.lineNumber isEqualTo  {
-                        select(max(orderLine2.lineNumber)) {
-                            from(orderLine2, "ol2")
-                            where { orderLine2.orderId isEqualTo orderLine.orderId }
-                        }
-                    }
-                }
-                orderBy(orderLine.orderId)
-            }
-
-            val expectedStatement = "select ol.order_id, ol.line_number " +
-                    "from OrderLine ol " +
-                    "where ol.line_number = " +
-                    "(select max(ol2.line_number) from OrderLine ol2 where ol2.order_id = ol.order_id) " +
-                    "order by order_id"
-
-            assertThat(selectStatement.selectStatement).isEqualTo(expectedStatement)
-
-            val rows = mapper.selectManyMappedRows(selectStatement)
-
-            assertThat(rows).hasSize(2)
-
-            assertThat(rows[0]).containsOnly(
-                entry("ORDER_ID", 1),
-                entry("LINE_NUMBER", 2)
-            )
-
-            assertThat(rows[1]).containsOnly(
-                entry("ORDER_ID", 2),
-                entry("LINE_NUMBER", 3)
-            )
-        }
     }
 }